Reorganize the `options` module
